*: merge AsyncPdClient into PdClient
